Fix target in makefile to synthesize with Vivado AES or ASCON asic implementation
**Makefile configuration for ASCON** :
ASCON_ARCH : default value is ascon_init. Select which architecture of ASCON to use, at the moment supported ascon_init which implements only the first round of Ascon and ascon_asip which is the architecture of Ascon generated with ASIP designer (https://www.synopsys.com/dw/ipdir.php?ds=asip-designer).
If selected  then the rtl can be configured with:
ASCON_SBOX_MODE : default value is lut. Select the implementation of the SBox to use, if the lut implemntation or the comb one. If selected combinational Sbox the only supported one is the standard.
ASCON_SBOX : default value is sbox_standard. The whole list of supported Sbox (only lut implementation) can be found in  and how the verilog defines are use in

**Makefile configuration for AES (NOT WORKING! needs to be implemented)** :
- AES_SBOX : defualt value is rijandael. The whole list of supported Sbox can be found in )
- AES_ARCH : defualt value is single_round. Selec the AES architectue to use, supported only the single round implementation *(NEEDED to support the configuration with an arbitrary length of rounds)*